04-analytics-engineering/pyproject.toml
[project]
name = "04-analytics-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "google-cloud-storage>=3.9.0",
]
